加了点判断

+++ goods/src/main/java/cn/edu/xmu/goods/controller/ShopController.java
@@ -4,6 +4,7 @@ import cn.edu.xmu.goods.model.bo.Shop;
 import cn.edu.xmu.goods.model.vo.*;
 import cn.edu.xmu.goods.service.ShopService;
 import cn.edu.xmu.ooad.util.Common;
+import cn.edu.xmu.ooad.util.ResponseCode;
 import cn.edu.xmu.ooad.util.ReturnObject;
 import io.swagger.annotations.*;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -66,11 +67,19 @@ public class ShopController {
      */
     @ApiOperation(value = "店家修改店铺信息", nickname = "modifyShop", notes = "", tags={ "shop", })
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "成功") })
+            @ApiResponse(code = 200, message = "成功"),
+            @ApiResponse(code = 140, message = "该店铺无法修改")})
     @PutMapping(value = "/shops/{id}")
     public Object modifyShop(Long id,Long shopId,@RequestBody ShopVo shopVo){
-        ReturnObject ret=shopService.updateShop(id,shopId,shopVo);
-        return Common.getNullRetObj(ret,httpServletResponse);
+        if(shopVo.getState()==6||shopVo.getState()==3)
+        {
+            return Common.decorateReturnObject(new ReturnObject(ResponseCode.MODIFYSHOP_ERROR));
+        }
+        else
+        {
+            ReturnObject ret=shopService.updateShop(id,shopId,shopVo);
+            return Common.getNullRetObj(ret,httpServletResponse);
+        }
     }
 
     /**
@@ -81,11 +90,20 @@ public class ShopController {
      */
     @ApiOperation(value = "管理员或店家关闭店铺", nickname = "deleteShop", notes = "如果店铺从未上线则物理删除",  tags={ "shop", })
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "成功") })
+            @ApiResponse(code = 200, message = "成功") ,
+            @ApiResponse(code = 180, message = "该店铺无法被执行关闭操作")
+    })
     @DeleteMapping(value = "/shops/{id}")
     public Object deleteShop(@ApiParam(value = "shop ID",required=true) @PathVariable("id") Long id){
-        ReturnObject ret=shopService.deleteShopById(id);
-        return Common.decorateReturnObject(ret);
+        if(shopService.getShopByShopId(id).getData().getState()==4)
+        {
+            ReturnObject ret=shopService.deleteShopById(id);
+            return Common.decorateReturnObject(ret);
+        }
+        else
+        {
+            return Common.decorateReturnObject(new ReturnObject(ResponseCode.SHUTSHOP_ERROR));
+        }
     }
 
     /**
@@ -96,11 +114,20 @@ public class ShopController {
      */
     @ApiOperation(value = "平台管理员审核店铺信息", nickname = "shopsShopIdNewshopsIdAuditPut", notes = "",  tags={ "shop", })
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "成功") })
+            @ApiResponse(code = 200, message = "成功") ,
+            @ApiResponse(code = 150, message = "该店铺不是待审核状态")
+    })
     @PutMapping(value = "/shops/{shopId}/newshops/{id}/audit")
     public Object shopsShopIdNewshopsIdAuditPut(@PathVariable("shopId") Long shopId,@ApiParam(value = "新店 ID",required=true) @PathVariable("id") Long id,@ApiParam(value = "" ,required=true )   @RequestBody ShopConclusionVo conclusion){
-        ReturnObject ret=shopService.passShop(id,conclusion);
-        return Shop.decorateReturnObject(ret);
+        if(shopService.getShopByShopId(id).getData().getState()==5)
+        {
+            ReturnObject ret=shopService.passShop(id,conclusion);
+            return Shop.decorateReturnObject(ret);
+        }
+        else
+        {
+            return Common.decorateReturnObject(new ReturnObject(ResponseCode.PASSSHOP_ERROR));
+        }
     }
 
 
@@ -112,11 +139,17 @@ public class ShopController {
      */
     @ApiOperation(value = "管理员上线店铺", nickname = "shopsIdOnshelvesPut", notes = "", tags={ "shop", })
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "成功") })
+            @ApiResponse(code = 200, message = "成功"),
+            @ApiResponse(code = 160, message = "该店铺无法上线")
+    })
     @PutMapping(value = "/shops/{id}/onshelves")
     public Object shopsIdOnshelvesPut(@PathVariable("id") long id){
-        ReturnObject ret=shopService.onShelfShop(id);
-        return Shop.decorateReturnObject(ret);
+        if(shopService.getShopByShopId(id).getData().getState()==5)
+        {
+            ReturnObject ret=shopService.onShelfShop(id);
+            return Shop.decorateReturnObject(ret);
+        }
+        else return Common.decorateReturnObject(new ReturnObject(ResponseCode.ONLINESHOP_ERROR));
     }
 
     /**
@@ -127,11 +160,16 @@ public class ShopController {
      */
     @ApiOperation(value = "管理员下线店铺", nickname = "shopsIdOffshelvesPut", notes = "", tags={ "shop", })
     @ApiResponses(value = {
-            @ApiResponse(code = 200, message = "成功") })
+            @ApiResponse(code = 200, message = "成功"),
+            @ApiResponse(code = 170, message = "该店铺无法下线")})
     @PutMapping(value = "/shops/{id}/offshelves")
     public Object shopsIdOffshelvesPut(@ApiParam(value = "店 ID",required=true) @PathVariable("id") long id){
-        ReturnObject ret=shopService.offShelfShop(id);
-        return Shop.decorateReturnObject(ret);
+        if(shopService.getShopByShopId(id).getData().getState()==4)
+        {
+            ReturnObject ret=shopService.offShelfShop(id);
+            return Shop.decorateReturnObject(ret);
+        }
+        else return Common.decorateReturnObject(new ReturnObject(ResponseCode.OFFLINESHOP_ERROR));
     }
 
 }
